+++ assets/src/ba_data/python/ba/ui/__init__.py
@@ -11,7 +11,6 @@ from typing import TYPE_CHECKING, cast, Type
 
 import _ba
 from ba._generated.enums import TimeType
-from ba._general import print_active_refs
 
 if TYPE_CHECKING:
     from typing import Any
@@ -230,10 +229,9 @@ def ui_upkeep() -> None:
                     'WARNING:',
                     obj,
                     'is still alive 5 second after its widget died;'
-                    ' you might have a memory leak.',
+                    ' you might have a memory leak. See efro.debug module'
+                    ' for tools to help debug this.',
                 )
-                print_active_refs(obj)
-
             else:
                 remainingchecks.append(check)
     ui.cleanupchecks = remainingchecks